Refactor to use epic's slab.
Previously this codebase had a fork of the 
ScalaNLP epic slab code.  This was primarily due
to there not being a release of the latest 
epic.  Now this code depends on a local build
of epic along with its Span.

The following classes should be removed.
org.sparkle.typesystem.Span
org.sparkle.slab.*